feat: remove forked process in favor or dispose
This will cause a hang if the endpoint still timesout.

BREAKING CHANGE: Refactor/remove the telemetry signatures. Remove the forked process. Requires a
