+++ packages/telemetry/src/exported.ts
@@ -2,4 +2,5 @@
 
 import { TelemetryReporter } from './telemetryReporter';
 
+export * from './telemetryReporter';
 export default TelemetryReporter;
