+++ packages/telemetry/test/unit/telemetryReporter.test.ts
@@ -1,294 +1,13 @@
-import set = Reflect.set;
-import { ConfigAggregator } from '@salesforce/core';
+import { ConfigAggregator, Logger } from '@salesforce/core';
 import { expect } from 'chai';
-import * as cp from 'child_process';
-import { describe, it } from 'mocha';
-import * as os from 'os';
 import * as sinon from 'sinon';
-import {
-  AppInsights,
-  buildPropertiesAndMeasurements,
-  getCpus,
-  getPlatformVersion,
-  TelemetryReporter
-} from '../../src/telemetryReporter';
-
-describe('AppInsights', () => {
-  const key = 'foo-bar-123';
-  const project = 'force-com-toolbelt';
-
-  let sandbox: sinon.SinonSandbox;
-  let trackStub: sinon.SinonStub;
-  let flushStub: sinon.SinonStub;
-  let osStub: sinon.SinonStub;
-  let configStub: sinon.SinonStub;
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox();
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-  });
-
-  it('should connect to app insights dashboard', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    const reporterKey = reporter.appInsightsClient ? reporter.appInsightsClient.config.instrumentationKey : null;
-    expect(reporter.appInsightsClient).to.not.be.undefined;
-    expect(reporterKey).to.equal(key);
-  });
-
-  it('should build base properties', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    const actualBaseProps = reporter.appInsightsClient ? reporter.appInsightsClient.commonProperties : {};
-    const expectedBaseProps = [
-      'common.cpus',
-      'common.os',
-      'common.platformversion',
-      'common.systemmemory',
-      'common.usertype'
-    ];
-    expect(Object.keys(actualBaseProps)).to.deep.equal(expectedBaseProps);
-  });
-
-  it('should add common properties', async () => {
-    const commonProperties = { foo: 'bar', hello: 'world' };
-    const options = { project, key, commonProperties };
-    const reporter = await AppInsights.create(options);
-    const actualProps = reporter.appInsightsClient ? reporter.appInsightsClient.commonProperties : {};
-    expect(actualProps).to.contain(commonProperties);
-
-    const actualPropsCount = Object.keys(actualProps).length;
-    const providedPropsCount = Object.keys(commonProperties).length;
-    expect(actualPropsCount).to.be.greaterThan(providedPropsCount);
-  });
-
-  it('should add context tags', async () => {
-    const contextTags = { foo: 'bar', hello: 'world' };
-    const options = { project, key, contextTags };
-    const reporter = await AppInsights.create(options);
-    const actualTags = reporter.appInsightsClient ? reporter.appInsightsClient.context.tags : {};
-    expect(actualTags).to.contain(contextTags);
-
-    const actualTagsCount = Object.keys(actualTags).length;
-    const providedTagsCount = Object.keys(contextTags).length;
-    expect(actualTagsCount).to.be.greaterThan(providedTagsCount);
-  });
-
-  it(`should replace GDPR sensitive value with ${AppInsights.GDPR_HIDDEN}`, async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    const actualTags = reporter.appInsightsClient ? reporter.appInsightsClient.context.tags : {};
-
-    expect(actualTags['ai.cloud.roleInstance']).to.equal(AppInsights.GDPR_HIDDEN);
-  });
-
-  it('should change url when using Asimov key', async () => {
-    const options = { project, key: 'AIF-12345' };
-    const reporter = await AppInsights.create(options);
-    const url = reporter.appInsightsClient ? reporter.appInsightsClient.config.endpointUrl : {};
-    expect(url).to.equal('https://vortex.data.microsoft.com/collect/v1');
-  });
-
-  it('should separate string attributes from numeric attributes', () => {
-    const attributes = { foo: 'bar', baz: 5 };
-    const expectedProperties = { foo: 'bar' };
-    const expectedMeasurements = { baz: 5 };
-    const { properties, measurements } = buildPropertiesAndMeasurements(attributes);
-    expect(properties).to.deep.equal(expectedProperties);
-    expect(measurements).to.deep.equal(expectedMeasurements);
-  });
-
-  it('should handle non existent appInsightsClient when sendTelemetryEvent is called', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    delete reporter.appInsightsClient;
-    expect(() => reporter.sendTelemetryEvent('testEvent'))
-      .to.throw(Error)
-      .and.have.property('name', 'sendFailed');
-  });
-
-  it('should handle non existent appInsightsClient when sendTelemetryException is called', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    delete reporter.appInsightsClient;
-    expect(() => reporter.sendTelemetryException(new Error('testException')))
-      .to.throw(Error)
-      .and.have.property('name', 'sendFailed');
-  });
-
-  it('should handle non existent appInsightsClient when sendTelemetryTrace is called', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    delete reporter.appInsightsClient;
-    expect(() => reporter.sendTelemetryTrace('testTrace'))
-      .to.throw(Error)
-      .and.have.property('name', 'sendFailed');
-  });
-
-  it('should handle non existent appInsightsClient when sendTelemetryMetric is called', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    delete reporter.appInsightsClient;
-    expect(() => reporter.sendTelemetryMetric('testMetric', 0))
-      .to.throw(Error)
-      .and.have.property('name', 'sendFailed');
-  });
-
-  it('should send telemetry event', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackEvent').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {});
-    }
-    reporter.sendTelemetryEvent('testEvent');
-    expect(trackStub.calledOnce).to.be.true;
-    expect(flushStub.calledOnce).to.be.true;
-  });
-
-  it('send telemetry event will fail unknown', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackEvent').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {
-        const error = new Error();
-        set(error, 'code', 'ExtraTerrestrialBiologicalEntityFromZetaReticuli');
-        throw error;
-      });
-    }
-    expect(() => {
-      reporter.sendTelemetryEvent('testEvent');
-    })
-      .to.throw(Error)
-      .and.have.property('name', 'unknownError');
-  });
-
-  it('should send telemetry exception', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackException').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {});
-    }
-    reporter.sendTelemetryException(new Error('testException'));
-    expect(trackStub.calledOnce).to.be.true;
-    expect(flushStub.calledOnce).to.be.true;
-  });
-
-  it('send telemetry exception will fail unknown', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackException').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {
-        const error = new Error();
-        set(error, 'code', 'ExtraTerrestrialBiologicalEntityFromZetaReticuli');
-        throw error;
-      });
-    }
-    expect(() => {
-      reporter.sendTelemetryException(new Error('testException'));
-    })
-      .to.throw(Error)
-      .and.have.property('name', 'unknownError');
-  });
-
-  it('should send telemetry trace', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackTrace').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {});
-    }
-    reporter.sendTelemetryTrace('testTrace');
-    expect(trackStub.calledOnce).to.be.true;
-    expect(flushStub.calledOnce).to.be.true;
-  });
-
-  it('send telemetry trace will fail unknown', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackTrace').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {
-        const error = new Error();
-        set(error, 'code', 'ExtraTerrestrialBiologicalEntityFromZetaReticuli');
-        throw error;
-      });
-    }
-    expect(() => {
-      reporter.sendTelemetryTrace('testTrace');
-    })
-      .to.throw(Error)
-      .and.have.property('name', 'unknownError');
-  });
-
-  it('should send telemetry metric', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackMetric').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {});
-    }
-    reporter.sendTelemetryMetric('testMetric', 0);
-    expect(trackStub.calledOnce).to.be.true;
-    expect(flushStub.calledOnce).to.be.true;
-  });
-
-  it('send telemetry metric will fail unknown', async () => {
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    if (reporter.appInsightsClient) {
-      trackStub = sandbox.stub(reporter.appInsightsClient, 'trackMetric').callsFake(() => {});
-      flushStub = sandbox.stub(reporter.appInsightsClient, 'flush').callsFake(() => {
-        const error = new Error();
-        set(error, 'code', 'ExtraTerrestrialBiologicalEntityFromZetaReticuli');
-        throw error;
-      });
-    }
-    expect(() => {
-      reporter.sendTelemetryMetric('testMetric', 0);
-    })
-      .to.throw(Error)
-      .and.have.property('name', 'unknownError');
-  });
-
-  it('should not create appInsightsClient when insights are disabled', async () => {
-    configStub = sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
-    const options = { project, key };
-    const reporter = await AppInsights.create(options);
-    expect(reporter.appInsightsClient).to.be.undefined;
-    expect(configStub.calledTwice).to.be.true;
-  });
-
-  it('should handle missing os.cpus value', () => {
-    osStub = sandbox.stub(os, 'cpus').callsFake(() => {});
-    const actual = getCpus();
-    expect(actual).to.equal('');
-    expect(osStub.calledOnce).to.be.true;
-  });
-
-  it('should handle missing os release value', () => {
-    osStub = sandbox.stub(os, 'release').callsFake(() => {});
-    const actual = getPlatformVersion();
-    expect(actual).to.equal('');
-    expect(osStub.calledOnce).to.be.true;
-  });
-});
+import { TelemetryReporter } from '../../src/telemetryReporter';
 
 describe('TelemetryReporter', () => {
   const key = 'foo-bar-123';
   const project = 'force-com-toolbelt';
 
   let sandbox: sinon.SinonSandbox;
-  let forkedProcessStub: sinon.SinonStub;
-  let killStub: sinon.SinonStub;
-  let sendStub: sinon.SinonStub;
-  let configStub: sinon.SinonStub;
 
   beforeEach(() => {
     sandbox = sinon.createSandbox();
@@ -298,83 +17,102 @@ describe('TelemetryReporter', () => {
     sandbox.restore();
   });
 
-  it('should created a forked process', async () => {
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({ kill: killStub });
-    const options = { project, key };
-    const reporter = await TelemetryReporter.create(options);
-    reporter.start();
-    expect(forkedProcessStub.calledTwice).to.be.true;
-  });
-
-  it('should kill a forked process after 3 seconds', async () => {
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({ kill: killStub });
-    const options = { project, key };
-    const reporter = await TelemetryReporter.create(options);
-    expect(reporter).to.not.be.undefined;
-    setTimeout(() => {
-      expect(killStub.calledOnce).to.be.true;
-    }, 3001);
-  });
-
   it('should send a telemetry event', async () => {
-    sendStub = sandbox.stub().callsFake(() => {});
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({
-      send: sendStub,
-      kill: killStub
-    });
     const options = { project, key };
     const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackEvent').callsFake(() => {});
+
     reporter.sendTelemetryEvent('testName');
     expect(sendStub.calledOnce).to.be.true;
   });
 
   it('should send a telemetry exception', async () => {
-    sendStub = sandbox.stub().callsFake(() => {});
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({
-      send: sendStub,
-      kill: killStub
-    });
     const options = { project, key };
     const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackException').callsFake(() => {});
+
     reporter.sendTelemetryException(new Error('testException'));
     expect(sendStub.calledOnce).to.be.true;
   });
 
   it('should send a telemetry trace', async () => {
-    sendStub = sandbox.stub().callsFake(() => {});
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({
-      send: sendStub,
-      kill: killStub
-    });
     const options = { project, key };
     const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackTrace').callsFake(() => {});
+
     reporter.sendTelemetryTrace('testTrace');
     expect(sendStub.calledOnce).to.be.true;
   });
 
   it('should send a telemetry metric', async () => {
-    sendStub = sandbox.stub().callsFake(() => {});
-    killStub = sandbox.stub().callsFake(() => {});
-    forkedProcessStub = sandbox.stub(cp, 'fork').returns({
-      send: sendStub,
-      kill: killStub
-    });
     const options = { project, key };
     const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackMetric').callsFake(() => {});
+
     reporter.sendTelemetryMetric('testMetric', 0);
     expect(sendStub.calledOnce).to.be.true;
   });
 
-  it('should not begin lifecycle when insights are disabled', async () => {
-    configStub = sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
+  it('should not send a telemetry event when disabled', async () => {
+    sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
     const options = { project, key };
     const reporter = await TelemetryReporter.create(options);
-    expect(reporter.forkedProcess).to.be.undefined;
-    expect(configStub.calledTwice).to.be.true;
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackEvent').callsFake(() => {});
+
+    reporter.sendTelemetryEvent('testName');
+    expect(sendStub.calledOnce).to.be.false;
+  });
+
+  it('should not send a telemetry exception when disabled', async () => {
+    sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
+    const options = { project, key };
+    const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackException').callsFake(() => {});
+
+    reporter.sendTelemetryException(new Error('testException'));
+    expect(sendStub.calledOnce).to.be.false;
+  });
+
+  it('should not send a telemetry trace when disabled', async () => {
+    sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
+    const options = { project, key };
+    const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackTrace').callsFake(() => {});
+
+    reporter.sendTelemetryTrace('testTrace');
+    expect(sendStub.calledOnce).to.be.false;
+  });
+
+  it('should not send a telemetry metric when disabled', async () => {
+    sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
+    const options = { project, key };
+    const reporter = await TelemetryReporter.create(options);
+    const sendStub = sandbox.stub(reporter.getTelemetryClient(), 'trackMetric').callsFake(() => {});
+
+    reporter.sendTelemetryMetric('testMetric', 0);
+    expect(sendStub.calledOnce).to.be.false;
+  });
+
+  it('should log to enable telemetry metric when disabled', async () => {
+    sandbox.stub(ConfigAggregator.prototype, 'getPropertyValue').returns(true);
+    const warn = sandbox.stub();
+    sandbox.stub(Logger, 'child').resolves({ warn, debug: sandbox.stub() });
+    const options = { project, key };
+    const reporter = await TelemetryReporter.create(options);
+
+    reporter.logTelemetryStatus();
+    expect(warn.calledOnce).to.be.true;
+    expect(warn.firstCall.args[0]).to.contain('=false');
+  });
+
+  it('should log to disable telemetry metric when enable', async () => {
+    const warn = sandbox.stub();
+    sandbox.stub(Logger, 'child').resolves({ warn, debug: sandbox.stub() });
+    const options = { project, key };
+    const reporter = await TelemetryReporter.create(options);
+
+    reporter.logTelemetryStatus();
+    expect(warn.calledOnce).to.be.true;
+    expect(warn.firstCall.args[0]).to.contain('=true');
   });
 });
